Plans 2024: Code hygiene/cleanup pre package extraction
