+++ client/landing/stepper/declarative-flow/internals/steps-repository/plans/plans-wrapper.tsx
@@ -33,7 +33,7 @@ import { setSelectedSiteId } from 'calypso/state/ui/actions';
 import { getSelectedSiteId } from 'calypso/state/ui/selectors';
 import { ONBOARD_STORE } from '../../../../stores';
 import type { OnboardSelect } from '@automattic/data-stores';
-import type { PlansIntent } from 'calypso/my-sites/plans-grid/hooks/npm-ready/data-store/use-grid-plans';
+import type { PlansIntent } from 'calypso/my-sites/plans-grid';
 import './style.scss';
 
 interface Props {
