+++ client/my-sites/plans-features-main/components/plan-upsell-modal/components/plan-upsell-button.tsx
@@ -1,5 +1,5 @@
 import { useTranslate } from 'i18n-calypso';
-import PlanButton from 'calypso/my-sites/plans-grid/components/plan-button';
+import { PlanButton } from 'calypso/my-sites/plans-grid';
 import type { PlanUpsellInfo } from '../hooks/use-plan-upsell-info';
 import type { PlanSlug } from '@automattic/calypso-products';
