+++ client/my-sites/plans-features-main/hooks/use-plan-intent-from-site-meta.ts
@@ -1,7 +1,7 @@
 import { useSiteIntent } from '@automattic/data-stores';
 import { useSelector } from 'react-redux';
 import { getSelectedSiteId } from 'calypso/state/ui/selectors';
-import type { PlansIntent } from 'calypso/my-sites/plans-grid/hooks/npm-ready/data-store/use-grid-plans';
+import type { PlansIntent } from 'calypso/my-sites/plans-grid';
 
 interface IntentFromSiteMeta {
 	processing: boolean;
