+++ client/my-sites/plans-grid/components/actions.tsx
@@ -22,9 +22,9 @@ import styled from '@emotion/styled';
 import { useSelect } from '@wordpress/data';
 import { useCallback } from '@wordpress/element';
 import { localize, TranslateResult, useTranslate } from 'i18n-calypso';
-import { useManageTooltipToggle } from 'calypso/my-sites/plans-grid/hooks/use-manage-tooltip-toggle';
 import { usePlansGridContext } from '../grid-context';
-import useDefaultStorageOption from '../hooks/npm-ready/data-store/use-default-storage-option';
+import useDefaultStorageOption from '../hooks/data-store/use-default-storage-option';
+import { useManageTooltipToggle } from '../hooks/use-manage-tooltip-toggle';
 import PlanButton from './plan-button';
 import { Plans2023Tooltip } from './plans-2023-tooltip';
 import type { PlanActionOverrides } from '../types';
