+++ client/my-sites/plans-grid/components/plan-features-container.tsx
@@ -1,11 +1,10 @@
 import { JetpackLogo } from '@automattic/components';
 import { LocalizeProps } from 'i18n-calypso';
-import { useManageTooltipToggle } from 'calypso/my-sites/plans-grid/hooks/use-manage-tooltip-toggle';
-import { DataResponse } from '../types';
+import { useManageTooltipToggle } from '../hooks/use-manage-tooltip-toggle';
+import { DataResponse, GridPlan } from '../types';
 import PlanFeatures2023GridFeatures from './features';
 import PlanDivOrTdContainer from './plan-div-td-container';
 import { Plans2023Tooltip } from './plans-2023-tooltip';
-import type { GridPlan } from '../hooks/npm-ready/data-store/use-grid-plans';
 
 const PlanFeaturesContainer: React.FC< {
 	plansWithFeatures: GridPlan[];
