+++ client/my-sites/plans-grid/components/plan-type-selector/components/interval-type-selector.tsx
@@ -1,4 +1,4 @@
-import { IntervalTypeProps } from '../types';
+import { IntervalTypeProps } from '../../../types';
 import { IntervalTypeDropdown } from './interval-type-dropdown';
 import { IntervalTypeToggle } from './interval-type-toggle';
