+++ client/my-sites/plans-grid/components/plan-type-selector/components/interval-type-toggle.tsx
@@ -6,9 +6,9 @@ import classNames from 'classnames';
 import { useTranslate } from 'i18n-calypso';
 import * as React from 'react';
 import useMaxDiscount from '../hooks/use-max-discount';
-import { IntervalTypeProps } from '../types';
 import generatePath from '../utils';
 import PopupMessages from './popup-messages';
+import type { IntervalTypeProps } from '../../../types';
 
 export const IntervalTypeToggle: React.FunctionComponent< IntervalTypeProps > = ( props ) => {
 	const translate = useTranslate();
