+++ client/my-sites/plans-grid/components/plan-type-selector/hooks/use-max-discount.ts
@@ -1,5 +1,5 @@
 import {
-	PlanSlug,
+	type PlanSlug,
 	TERM_ANNUALLY,
 	getPlanSlugForTermVariant,
 	isMonthly,
