+++ client/my-sites/plans-grid/components/plan-type-selector/style.scss
@@ -1,4 +1,4 @@
-@import "calypso/my-sites/plans-grid/media-queries";
+@import "../../media-queries";
 
 .plan-type-selector {
 	width: 100%;
