+++ client/my-sites/plans-grid/components/popular-badge.tsx
@@ -1,7 +1,7 @@
 import { getPlanClass, PlanSlug } from '@automattic/calypso-products';
 import classNames from 'classnames';
-import PlanPill from 'calypso/my-sites/plans-grid/components/shared/plan-pill';
 import { usePlansGridContext } from '../grid-context';
+import PlanPill from './shared/plan-pill';
 
 const PopularBadge: React.FunctionComponent< {
 	isInSignup?: boolean;
