+++ client/my-sites/plans-grid/components/storage-add-on-dropdown.tsx
@@ -4,8 +4,8 @@ import { useDispatch, useSelect } from '@wordpress/data';
 import { useCallback, useEffect, useMemo } from '@wordpress/element';
 import { useTranslate } from 'i18n-calypso';
 import { usePlansGridContext } from '../grid-context';
-import useDefaultStorageOption from '../hooks/npm-ready/data-store/use-default-storage-option';
-import useIsLargeCurrency from '../hooks/npm-ready/use-is-large-currency';
+import useDefaultStorageOption from '../hooks/data-store/use-default-storage-option';
+import useIsLargeCurrency from '../hooks/use-is-large-currency';
 import { getStorageStringFromFeature } from '../util';
 import type { PlanSlug, StorageOption, WPComStorageAddOnSlug } from '@automattic/calypso-products';
 import type { AddOnMeta } from '@automattic/data-stores';
