+++ client/my-sites/plans-grid/grid-context.tsx
@@ -1,6 +1,5 @@
 import { createContext, useContext } from '@wordpress/element';
-import type { GridPlan, PlansIntent } from './hooks/npm-ready/data-store/use-grid-plans';
-import type { GridContextProps } from './types';
+import type { GridContextProps, GridPlan, PlansIntent } from './types';
 import type { FeatureList } from '@automattic/calypso-products';
 import type { Plans } from '@automattic/data-stores';
