+++ client/my-sites/plans-grid/hooks/data-store/use-highlight-labels.ts
@@ -6,9 +6,9 @@ import {
 	type PlanSlug,
 } from '@automattic/calypso-products';
 import { useTranslate } from 'i18n-calypso';
-import { isSamePlan } from '../../../lib/is-same-plan';
+import { isSamePlan } from '../../lib/is-same-plan';
 import { isPopularPlan } from './is-popular-plan';
-import type { PlansIntent } from './use-grid-plans';
+import type { PlansIntent } from '../../types';
 import type { TranslateResult } from 'i18n-calypso';
 
 interface Props {
