+++ client/my-sites/plans-grid/lib/get-plan-pricing-info-from-grid-plans.ts
@@ -1,4 +1,4 @@
-import type { GridPlan } from 'calypso/my-sites/plans-grid/hooks/npm-ready/data-store/use-grid-plans';
+import type { GridPlan } from '../types';
 
 type PricingInfo = {
 	prices: number[];
