+++ client/my-sites/plans-grid/style.scss
@@ -1,5 +1,5 @@
 @import "@automattic/onboarding/styles/mixins";
-@import "calypso/my-sites/plans-grid/media-queries";
+@import "./media-queries";
 
 $table-cell-max-width: 344px;
 $mobile-card-max-width: 440px;
